Addressed a number of PR comments
Please see: https://github.com/appsembler/edx-figures/pull/1

NOTE: This is a work in progress (WIP) with the next stage to
incorporate UI work that Matej is working on. So the asset management is
going to likely change

I pulled out installation and configuration instructions from the README
since there were significant comments in the PR about this and this is
definitely going to change as we work to make this package easier to
intergrate with edx-platform.