Added devsite and core testing infrastructure
devsite has two settings files, 'settings.py' to run a barebones
development server to test the front end and 'test_settings.py' to be
used with the automated tests

Basic test infrastructure tests the CourseOverviewFilter. This commit is
intended to provide the bare essentials to do some 'real testing'

Also included is a 'test_sanity_check' test that asserts True. The
purpose of this test is to run it to make sure the project
infrastructure is set up. We may just remove it, but it was helpful in
initial development