Updated pipeline tasks
Added an experimental feature that runs the course metrics populators in
parallel, then the site metrics populator after all the courses are
done. You can run this via the `populate_figures_metrics` Django
management command using the `--experimental` flag.

WARNING: When using this experimental feature, the job tends to get
stuck. We're working on addressing it. This is why it is 'experimental'

NOTICE: The experimental job also excludes courses with greater than 50
enrollments, this is to speed up live testing in devstack when there are
courses with many enrollments in your devstack data.

* Removed CourseDailyMetricsJob as it is not needed
* Created individual task functions in `figures/tasks.py`. This will
both help development by being able to run in Celery individual course
ETL jobs. They also serve to run subtasks for the experimental job top
level task