Fixed SiteDailyMetrics prior day data accumulation
This commit addresses the problem that if the immediate prior day record
does not exist for SiteDailyMetrics, then the cumulative daily active
site user count uses zero '0' as the previous day.

What it does now is retrieve the latest prior SiteDailyMetrics record
before the given date. This given date defaults to 'today', but can be
set to any date

NOTE: This does NOT fix the issue that the SiteDailyMetrics active user
count has flawed data. The flaw is that it collects the totals from the
CourseDailyMetrics records for the site for the matching day. The
problem with this is that it will count users multiple times. For
example, if a user is active in 2 different courses, then the user will
get counted once for each course when added to the site cumulative
active users instead of just once. This bug still needs to be addressed