Fixed LearnerCourseGradesDetailsSerializer.get_progress_data method
If there aren't any matching LearnerCourseGradeMetrics records matching
the filtering criteria then an exception is raised and trapped,
triggering an error log.

This does not change the behvior of the method. What is does is reduce
logging noise to the log files and also does not create a
`PipelineError` table log enty, saving database space