Added new functions to `figures.sites` to support query abstractions
- Added `course_enrollments_for_course`
- Added `student_modules_for_course_enrollment`

This is to support the course progress pipeline performance improvement
code restructuring. The purpose of adding these functions is to move
forward with using `figures.sites` as the source of platform data. To
note, there are a number of places in the code that do not follow this.
As we rework Figures, we will refactor code to use `figures.sites`
instead of directly querying the platform models.

Note the function name change from the existing functions. We are moving
away from using the `get_` prefix for the site functions.