Fixes incorrect site monthly metrics course completion data
Site monthly metrics course completions API endpoint was retrieving
total site users erroneously. This was caused by 'copy/paste of the
viewset method and the mismatch never got caught. This is a good reason
why testing data quality on top of basic code coverage is important for
the API views

* https://appsembler.atlassian.net/browse/RED-1061