progress indicator for populate_daily_metrics
This is a simple quality of life improvement.

The daily pipeline takes a long time to run, and currently gives no
indication of how far along in the process it is.

Adding a log message like this means that I can quickly tail the logs
and form a ballpark estimate of how much longer it will be
running. Eg, if we want to deploy (which we need to stop the pipeline
for currently), seeing that it's at 1995 of 2000, I'd be inclined to
just wait a few minutes for it to finish rather than stop the
pipeline, deploy, and restart it.

Similarly, if it's been running for 12 hours and we see that it's only
made it 10% of the way through, that would indicate that there's a bug
or some issue slowing it down and we need to stop the pipeline and
investigate before the next day's pipeline starts running and things
get even worse.