Added 'mau' field to 'SiteDailyMetrics' model
The purpose of this is to improve API performance. Figures currently
performs a live query on the `edx-platform`
`courseware.models.StudentModule` model, but the query is too slow when
there are many `StudentModule` records. The rough order of magnitude
(ROM) for "many" appears to be about 10^6.

The slow query causes an API timeout, degrading performance for Figures
UI.

This commit provides the storage for this metric so the following commit
can update the pipeline code and metrics code