Added 'mau' collection to Figures pipeline
The purpose of this is to improve API performance. See the previous
commit with SHA 92ea8c460d3ab3ae7cc42945f94f27c218dde39f for an
explanation.

This commit adds the pipeline code to populate 'SiteDailyMetrics.mau'
with the latest numbers for the month as of yesterday

We also clean up the old live MAU collection from StudentModule,
removing dead coe and the tests that call the dead code