Enrollment Metrics View test - improvments
Added test coverage for when viewset does not pagingate. This can be a
case where the existing class paginagor is removed or set to None for
the class being inherited.

Instead of doing this, we could have just assumed that the results are
always paginated in the EnrollmentMetricsViewSet completed and
comletede_ids actions. However, the example I learned from did the conditional
check for paginated results.
So this is a bit of cheap defensive programming to allow the viewset
method to paginate or not and have test coverage for this