Added execution time context manager logger
Added `figures.log.log_exec_time` as a context manager to output the log
time for a block of code executed within.

The purpose of this is as a convenience to logging execution time. At
this time, it does not handle variables within the block so it only
outputs the label passed in as a parameter