Fix certificates count as proxy for course completions
Reworked function to get the latest sum of cerfificates across courses
for the site. This is done by querying CourseDailyMetrics.

* Added new test module to focus on testing certificate metrics

This is a workaround. It fixes the metric, but is really just a patch. What needs to happen is to rethink this metric, what we want to say and how to efficiently (storage and query performance) manage certificates achived over time.

Especially since we only want to capture data as they change