Improved query performance for LearnerMetricsViewSet
* Added prefetch_related to the User model in LearnerMetricsViewSet to
query the related "CourseEnrollment" records
* Updated test coverage to include users, courses, and enrollments for
two sites in order to ensure we are retrieving only the data we want and
avoid "site bleed". For this, created an initial generalized fixture in
"tests/conftest.py". This initial fixture contains User, CourseOverview,
CourseEnrollment, Organization, and Site records. If the version of
"organizations" is our fork (supports multi-site) then
OrganizationCourse and UserOrganizationMapping records are also filled

This commit lays the groundwork for additional performance imporovement
to reduce the number of queries needed in the serializers used for
LearnerMetricsViewSet