Learner Metrics Serializer - Fixed bug
There were missing tests. We checked for each learner, but didn't
check for each learner's enrollments. What was happening with the
`prefetch_related` is that it can't filter just the enrollments we want
in the related table (or at least I haven't found a way to do this yet)
so although we filtered on the learners, we weren't able to filter on
the learner enrollments. So now we need to filter the enrollments for
each learner against the course id list.

Now, we don't strictly need to do this. The front end will filter out
unwanted courses. However, when the query params say "I want these
courses", then it is good form to just return those courses. If we run
into performance issues, then we can return all courses to reduce the
query count and come up with an improved solution