Reworked SiteMonthlyMetrics registered users metric
*  now use collected total user counts in
SiteDailyMetrics.total_user_count for the registered users
* This saves expensive database queries
* This is more accurate because it preserves historical data. The
previous approach captured live data for users for a site and users can
be removed from sites
* Very rudimentary test coverage added

What can be improved:

This does not report on users who have enrolled
for the current day. We want to avoide querying the Django 'User' model
for 'date_joined' because the field is not indexed

Test coverage can be improved. This commit just includes very basic
coverage