Fixed Learner Metrics API endpoint duplicate users returned
When the learner-metrics API endpoint was called with a user filter,
such as username, email, name, duplicate entries would be returned. the
count would match the number of enrollments the learner has.

This fix adds the `distinct=True` option to the char filters for the
UserFilterSet and appears to address the issue.

This was happening because Django Filter does not default to distinct
records when filtering with `CharFilter`.

Some basic testing added for this, but we should have futher testing on
all the query parameters for `figures.filters.UserFilterSet`, possibly
creating a new filterset object specific to the learner metrics API
endpoint