Figures devsite.seed - new functions
1. Added `seed_course_enrollments_multiple` - This creates a set of
users and creates enrollments for each user for every course in the
site. The driver to make this is to test user filtering because the
learner-metrics endpoint (using the UserFilterSet) was returning
duplicates. It turns out that you need to specify 'distinct' in the
filter class constructor. See here: https://github.com/appsembler/figures/pull/273

2. Added `add_users_to_site` as a helper to create
`UserOrganizationMapping` records for a set of users to a given site