Fix Figures devsite settings
Enabled setting OPENEDX_RELEASE frrom the `.env` file. This allows the
developer to set the Open edX release version via the `.env` file with
an override from a shell env var

This only updates the devsite settings and not the test settings. We'll
follow on with updating the test settings in Figures 0.4