Update devsite requirements files to support signals
We need to get the request object to get the site identifier. This
commit adds `crum` to the requirements files for that

Also added `edx-django-sites-extensions` to help Figures devsite act
more like the platform for multisite behavior. But we're not all the way
there yet, this commit just includes the needed package