Use the Open edX named releases to conditionally import platform objects
This commit exists to address the problem that a RuntimeError can be
raised instead of an ImportError when importing objects from
ed-platform.

The options to fix are check for ImportError AND RuntimeError. However,
we opt (at least for now) to check the Open edX named release to
conditionally import objects when the namespace has changed between
named releases.

Here is an example of the error when only ImportError is handled:
```
RuntimeError: Conflicting 'studentmodule' models in application 'courseware': <class 'courseware.models.StudentModule'> and <class 'lms.djangoapps.courseware.models.StudentModule'>.
```

NOTE: This may break Juniper compat as it was not tested in a live
Juniper devstack

This commit also addresses https://github.com/appsembler/figures/issues/290

We are missing test coverage for figures.compat.course_grade, but this
requires some thinking because for the ginkgo test case, we need to mock
all the imports as we run this test agnostically (this may be the only
test that doesn't or is very close to not relying on figures 'mocks'
test infrastructure. To test 'course_grade' we need to check that
CourseGradeFactory().create(...) is called for Ginkgo, and 'read' is
called for everyone else. BUT we are already mocking CourseGradeFactory
so the import doesn't fail and attempts I made so far to get
"course_grade_factory_mock.created.assert_called()" have failed. So I
ran out of time and punting until later to get test coverage here

Also, we've mashed up a chain of mocks for three tests. We are
effectively running each test to run each of the supported Open edX
platform releases: Ginkgo, Hawthorn, Juniper.

This could use some rethinking...