Add db indexes to SiteDailyMetrics and CourseDailyMetrics
This commit adds an index to the SiteDailyMetrics field: 'date_for'

This commit adds indexes to the CourseDailyMetrics fields: 'course_id'
and 'date_for'

This commit adds a new migration, '0014' to perform non-atomic separate
state and database migration actions. Included is compatibility
functionality so that the migration works for both MySQL and SQLite.