0.4 LPO performance update - figures.compat
This is some groundwork to improving the "learner-metrics" API endpoint.

Added new function, 'course_grade_from_course_id' to figures.compat to
simplify and abstract retrieving the course grade structure from
CoureGradeFactory.

Also started with moving edx-platform imports into compat from other
modules to centralize the dependency on the platform to a single module.
The goal is for figures.compat to serve as an adapter to direct access
to platform code