0.4 LPO performance update - new EnrollmentProgress class
Added a new package, 'figures.progress' instead of adding to
'figures.metrics' in order to prevent cyclical dependencies causing
errors in figures.models

This is a stripped down version of figures.metrics.LearnerCourseGrades
class. The purpose is to minimize queries and object construction in
order to get course structure data
0.4 LPO performance update - new EnrollmentProgress class

Added a new package, 'figures.progress' instead of adding to
'figures.metrics' in order to prevent cyclical dependencies causing
errors in figures.models

This is a stripped down version of figures.metrics.LearnerCourseGrades
class. The purpose is to minimize queries and object construction in
order to get course structure data
0.4 LPO performance update - new EnrollmentProgress class

Added a new package, 'figures.progress' instead of adding to
'figures.metrics' in order to prevent cyclical dependencies causing
errors in figures.models

This is a stripped down version of figures.metrics.LearnerCourseGrades
class. The purpose is to minimize queries and object construction in
order to get course structure data
0.4 LPO performance update - new EnrollmentProgress class

Added a new package, 'figures.progress' instead of adding to
'figures.metrics' in order to prevent cyclical dependencies causing
errors in figures.models

This is a stripped down version of figures.metrics.LearnerCourseGrades
class. The purpose is to minimize queries and object construction in
order to get course structure data