0.4 LPO performance update - New EnrollmentData model
* This commit adds a new model to Figures, "EnrollmentData"
* This commit is a foundational piece to improve the learner progress
overview page performance

The purpose of this is to allow learner metrics data to be retrieved
with a minimum of queries. This model collects a snapshot of a learner's
progress for a given course (an "enrollment"). It forms the foundation
to speed up the 'learner-metrics' API endpoint by reducing the queries
needed to collect data. Prior to this work, 'learner-metrics' needed to
query the platform's CourseEnrollment model and find the most recent
instance of Figures LearnerCourseGradeMetrics model (if it exists: The
learner needs to make course progress for it to exist) for the course.
All this is quite expensive.

Included in this commit is the model, migration, admin update,
EnrollmentDataFactory for tests and the most barebones of test cases
specifically for the model.

Testing needs to be expanded. Read the module docstring in
'test_enrollment_data_model' for specific tests we should construct