0.4 LPO performance update - Updated views and serializers
Added new viewset and new serializers for version 2 of the
'learner-metrics' endpoint

We keep the old viewset and serializers (for now) to test data
correctness with a remote client (which I will write) and to benchmark
relative performance differences.

The old viewset is moved to endpoint `/figures/api/learner-metrics-v1`

The new viewset uses the existing endpoint. This is so we can test the
front end against the new endpoint

Renamed the test module for v1 of the viewset

Added a new test module for the v2 of the viewset, but we need to update
the test data to add mock EnrollmentData objects as the tests now fail,
so the test module has 'pytest.skip' decoration.