0.4 LPO performance update - Add backfill exception handling
In the figures.backfill.backfill_enrollment_data_for_site function:

If an edx-platform `course` structure cannot be found from the course
id, the error will be trapped and added to the 'errors' key in the
results. The purpose of this is to not fail the whole site processing if
there are defects for a single course.

This can happen if there is an "unlinked CourseEnrollment" where the
course content was deleted or the course key changed. This should not
happen in production, but does happen on staging, so we handle it