Fix tox.ini issue and minor changes
The Hawthorn tests were not running due to a recent typo I added in the
tox.ini 'commands' section. Fixed that.

Also calling out Hawthorn in the same way as Ginko and Juniper. Now all
have Open edX release specific pytest-<named release>.ini files and
Hawthorn now specifies the 'OPENEDX_RELEASE' also

This means that regardless of what the default environment that is set
in the settings, the Hawthorn tests will specify the proper environment