Figures sites and pipeline fixes
Fixed Figures sites - Some functions would not work in standalone

Fixed date/datetime type in site and course daily metrics pipeline
loaders. The problem was that the loaders would filter for existing
records using a datetime where the 'date_for' field is a date field.
Using a datetime caused the date to check the previous day in the
queryset filter

Fixed logic in the enrollment metrics pipeline. The enrollment metrics
pipeline code was causing massive noise in the logs. The key change is
in working with the StudentModule records. Records are filtered now per
enrollment instead of per course and if there are no StudentModule
records then the call returns back to the caller without running a query
for the LearnerCourseGradeMetrics records for th enrollment

Reworked the logging in enrollment_metrics so that if an LCGM record is
found for the enrollment without any StudentModule records, then a
warning is logged with the specific prefix 'FIGURES:PIPELINE:LCGM' and
the LCGM record id as well as the user id and course id. This will help
with easily grepping the log files to aid in safe cleanup of LCGM
records without deleting history.

While reworking the logging, also changed code to only log to the Python
logger instead of Figures 'PipelineError' model as we overhaul logging

Added some light test coverage for the changes, but still needs more