Figures pipeline fixes and improvements
Refactored functionality - pipeline rule for date handling

* Added function, 'pipeline_date_for_rule' in the new module:
  figures.pipeline.helpers

* This refactoring extracts 'date for' logic out of the daily metrics modules
into a common helper module and function. Read the docs in
`figures.pipeline.helpers`. The logic isn't hard but a bit brittle as
this refactoring addresses a bug where using a datetime to query the
daily metrics models' 'date_for' field caused the previous day's record
to be retrieved. Because getting the date handling correct is critical
to accurate metrics, we've given it some special attention here by
pulling the duplicate code out of both the site and course daily metrics
data loaders into a common library function

* Added code coverage for the new pipeline helper function. Even though
the functionality is pretty mundane, it is a key piece to pulling
correct data, so the tests should cover testing correctness and type
handling