Reorganize the requirements files for Juniper tox tests
The purpose of this commit is to clean up the requirements files. This
makes things a bit clearer. We now have a "Juniper community" mode and a
"Juniper multisite" mode.

* We remove Py38 because Juniper needs to run in Py35
* The `juniper_community.txt` and `juniper_multisite.txt` requirements
files read from `juniper_base.txt`. Each of the 'community' and
'multisite' have only the package differences, which at this point is
the version/branch of `edx-organizations`