build and publish figures on pypi
This has been made from the Figures build and release steps in our confluence docs.

This is still in testing and it's good not to mess up the Figures PyPI so this action will only publish `.dev` releases as described in the setuptools of Python: https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#publishing-to-package-registries

One this is tested and working, we'll publish for all releases.

This will enable a GitHub-based workflow which is completely different from the previous command line based one:

 - Update the version in the `setup.py` and make a PR
 - Merge the release bump PR
 - Publish a release on GitHub (done via the GUI)
 - The release is published automatically on PyPI

The workflow above is different from our [PyPI release Runbook](https://appsembler.atlassian.net/wiki/spaces/ED/pages/41550165/New+PyPI+release).