Add stale course enrollment handling to pipeline
What's included here is a generator to yield CourseEnrollment records
that need to have their EnrollmentData records updated

If we install Figures on a site that's already been running or we  miss a day
or more in processing, then the daily processing will not pick up on all
"stale enrollments", enrollments that Figures EnrollmentData record updated.

This is because Figures daily metrics jobs only look at "yesterday" in
order to work around the limitations of StudentModule.

The real logic is in `figures.enrollment.is_enrollment_data_out_of_date`
(Which is still pretty simple)