Adds daily metrics backfill functions
This commit has new functions in `figures.pipeline.backfill`.
For each function, see the details in each function's docstring

Because we don't want to update courses created after the "date_for", we
need to identify which courses to update. That's what the bulk of this
commit's code does

It uses the existing metrics loader classes to do the actual data
querying and metrics model saves

Add backfill dialy metrics managmenet command

* Still working on getting the mgmt test to pass. Having errors in parsing
the args call to `call_command`, Will pick up again to get this working
tomorrow